% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{The Max-Min Normalization}
\usage{
normalization(x, margin = 2)
}
\arguments{
\item{x}{the dataset.}

\item{margin}{data is normalized by row (margin = 1) or by column (margin = 2). The default is 2.}
}
\description{
This function normalizes the data using the max-min normalization
}
\examples{

data(ozone)
scaled.data = normalization(ozone[,-1])
ozone.scale = data.frame(y = as.character(ozone[,1]), scaled.data[,-1])

}
\author{
Wanwan Zheng
}
